Исправила: в users удалила параметр для email_verifid; в icons для id вернула increments; в wallets исправила метод down; для icon-id исправила во всех таблицах тип integer

+++ database/migrations/2020_07_06_192311_create_incomes_table.php
@@ -19,7 +19,7 @@ class CreateIncomesTable extends Migration
             $table->string('name', 45);
             $table->decimal('amount', 10,2)->default(0);
             $table->string('description', 255)->nullable();
-            $table->bigInteger('icon_id')->unsigned();
+            $table->integer('icon_id')->unsigned();
         });
 
         Schema::table('incomes', function (Blueprint $table){
